+++ packages/@tinacms/graphql/src/schema/validate.ts
@@ -1,7 +1,7 @@
 import { addNamespaceToSchema } from '@tinacms/schema-tools';
-import deepClone from 'lodash.clonedeep';
 import { sequential } from '../util';
 import * as yup from 'yup';
+import { cloneDeep } from 'es-toolkit';
 
 import {
   type TinaField,
@@ -25,7 +25,7 @@ const FIELD_TYPES: TinaField['type'][] = [
 
 export const validateSchema = async (schema: Schema) => {
   const schema2: Schema<true> = addNamespaceToSchema<Schema<true>>(
-    deepClone(schema) as unknown as Schema<true>
+    cloneDeep(schema) as unknown as Schema<true>
   );
   const collections = await sequential(
     schema2.collections,