+++ packages/@tinacms/search/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tinacms/search
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`f2577b9`](https://github.com/tinacms/tinacms/commit/f2577b911a97ecc1c3f53a98ae8218cc33bc9867), [`38920ce`](https://github.com/tinacms/tinacms/commit/38920ce29a9b63c54b04f39537f19beab62d2c86), [`d1bd1a1`](https://github.com/tinacms/tinacms/commit/d1bd1a1312fa910e237ec06608e7c11830c78346), [`4b824be`](https://github.com/tinacms/tinacms/commit/4b824be53572f9231753ebd3b5f14fd778fd73d6)]:
+  - @tinacms/schema-tools@1.10.0
+  - @tinacms/graphql@1.6.2
+
 ## 1.1.1
 
 ### Patch Changes